Fix a bug in client/ct on 32-bit systems.
Exhibited itself as a "STH has bad timestamp" error. See this thread for more
information:

https://groups.google.com/d/topic/certificate-transparency/Qlp7avuDzLM/discussion

+++ cpp/client/ct.cc
@@ -791,7 +791,7 @@ int GetSTH() {
   LogVerifier *verifier = GetLogVerifierFromFlags();
 
   // Allow for 10 seconds of clock skew
-  uint64_t latest = (time(NULL) + 10) * 1000;
+  uint64_t latest = ((uint64_t) time(NULL) + 10) * 1000;
   LogVerifier::VerifyResult result
       = verifier->VerifySignedTreeHead(sth, 0, latest);
